+++ frontend/src/components/header/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Header";
