+++ frontend/src/components/img-preview/ImagePreview.js
@@ -1,9 +1,8 @@
 import Card from "@material-ui/core/Card";
 import CardMedia from "@material-ui/core/CardMedia";
 import { makeStyles } from "@material-ui/core/styles";
-
-import React from "react";
 import Skeleton from "@material-ui/lab/Skeleton/Skeleton";
+import React from "react";
 
 const useStyles = makeStyles((theme) => ({
   card: {
