+++ frontend/src/components/img-preview/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ImagePreview";
