+++ frontend/src/components/index.js
@@ -1,0 +1,6 @@
+export { default as Header } from "./header";
+export { default as ImagePreview } from "./img-preview";
+export { default as withPageContainer } from "./page-containter";
+export { default as UploadButton } from "./upload-button";
+export { default as UploadForm } from "./upload-form";
+export { default as Theme } from "./theme";
