+++ frontend/src/components/upload-button/index.js
@@ -1,0 +1,1 @@
+export { default } from "./UploadButton";
