+++ frontend/src/components/upload-form/index.js
@@ -1,0 +1,1 @@
+export { default } from "./UploadForm";
