+++ frontend/src/pages/about/index.js
@@ -1,0 +1,1 @@
+export { default } from "./About";
