+++ frontend/src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as About } from "./about";
+export { default as Decode } from "./decode";
+export { default as Encode } from "./encode";
+export { default as Team } from "./team";
