+++ frontend/src/pages/team/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Team";
